+++ pip/vcs/bazaar.py
@@ -1,10 +1,9 @@
 import os
 import tempfile
 import re
-from pip import call_subprocess
 from pip.backwardcompat import urlparse
 from pip.log import logger
-from pip.util import rmtree, display_path
+from pip.util import rmtree, display_path, call_subprocess
 from pip.vcs import vcs, VersionControl
 from pip.download import path_to_url2
