+++ pip/vcs/git.py
@@ -1,6 +1,6 @@
 import tempfile
 import re
-from pip import call_subprocess
+from pip.util import call_subprocess
 from pip.util import display_path, rmtree
 from pip.vcs import vcs, VersionControl
 from pip.log import logger
